// src/app/shared/models/weather.ts
export interface WeatherModel {
    date?: string;
    time?: string;
    weather?: string;
    icon?: string;
    temperature?: number;
    tempMin?: number;
    tempMax?: number;
    humidity?: number;
    pressure?: number;
    wind?: string;
    sunrise?: string;
    sunset?: string;
    dayTime?: string;
    lat?: number,
    lon?: number,
}